services/llm-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.7.4
httpx==0.27.0
PyMuPDF==1.24.5 # For reading pdfs directly (fitz)
google-genai==0.4.0
anthropic==0.28.1
Pillow==10.3.0
pytesseract==0.3.10
cryptography==42.0.4
openai==1.28.1
